+++ data/RankColors.lua
@@ -1,0 +1,7 @@
+RankColors = {
+    ["S"] = "ffff8000", -- Legendary
+    ["A"] = "ffa335ee", -- Epic
+    ["B"] = "ff0070dd", -- Rare
+    ["C"] = "ff1eff00", -- Uncommon
+    ["D"] = "ff9d9d9d", -- Poor
+}