+++ data/SpecTierlIst.lua
@@ -1,0 +1,67 @@
+SpecTierlist = {
+    -- Death Knight
+    [250] = "B", -- Blood
+    [251] = "B", -- Frost
+    [252] = "S", -- Unholy
+
+    -- Demon Hunter
+    [577] = "B", -- Havoc
+    [581] = "B", -- Vengeance
+    [1480] = "S", -- Devourer
+
+    -- Druid
+    [102] = "B", -- Balance
+    [103] = "A", -- Feral
+    [104] = "A", -- Guardian
+    [105] = "S", -- Restoration
+
+    -- Evoker
+    [1467] = "C", -- Devastation
+    [1468] = "A", -- Preservation
+    [1473] = "S", -- Augmentation
+
+    -- Hunter
+    [253] = "B", -- Beast Mastery
+    [254] = "B", -- Marksmanship
+    [255] = "A", -- Survival
+
+    -- Mage
+    [62] = "B", -- Arcane
+    [63] = "C", -- Fire
+    [64] = "B", -- Frost
+
+    -- Monk
+    [268] = "S", -- Brewmaster
+    [269] = "B", -- Windwalker
+    [270] = "S", -- Mistweaver
+
+    -- Paladin
+    [65] = "A", -- Holy
+    [66] = "C", -- Protection
+    [70] = "A", -- Retribution
+
+    -- Priest
+    [256] = "S", -- Discipline
+    [257] = "B", -- Holy
+    [258] = "A", -- Shadow
+
+    -- Rogue
+    [259] = "A", -- Assassination
+    [260] = "A", -- Outlaw
+    [261] = "B", -- Subtlety
+
+    -- Shaman
+    [262] = "A", -- Elemental
+    [263] = "B", -- Enhancement
+    [264] = "S", -- Restoration
+
+    -- Warlock
+    [265] = "B", -- Affliction
+    [266] = "S", -- Demonology
+    [267] = "C", -- Destruction
+
+    -- Warrior
+    [71] = "S", -- Arms
+    [72] = "A", -- Fury
+    [73] = "B", -- Protection
+}